Moved Scene View Picking functionality to its own static class.
This way you can easily add scene view picking functionality to custom property drawers.
