+++ Scripts/Editor/SceneViewPicking.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae7c4a848c0dc9144a2050634385c700
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
